change some classes and methods

+++ Task1/Task1/Classes/Caramel.cs
@@ -10,15 +10,16 @@ namespace Task1.Classes
     class Caramel : ISweet
     {
         private const double CARAMEL_PRICE = 4.20;
+
         public string Name => throw new NotImplementedException();
 
-        public int Weight => throw new NotImplementedException();
+        public double Weight => throw new NotImplementedException();
 
-        public int Sugar => throw new NotImplementedException();
+        public double Sugar => throw new NotImplementedException();
 
         public double Calories => throw new NotImplementedException();
 
-        public Caramel(string name , int weight , int sugar, double calories)
+        public Caramel(string name , double weight , double sugar, double calories)
         {
             name = Name;
             weight = Weight;
