+++ Task1/Task1/Classes/ChocolateBar.cs
@@ -13,9 +13,9 @@ namespace Task1.Classes
 
         public string Name => throw new NotImplementedException();
 
-        public int Weight => throw new NotImplementedException();
+        public double Weight => throw new NotImplementedException();
 
-        public int Sugar => throw new NotImplementedException();
+        public double Sugar => throw new NotImplementedException();
 
         public double Calories => throw new NotImplementedException();
